Fix ptr unsigned error for cygwin
git-svn-id: svn://svn.macaulay2.com/Macaulay2/branches/gfurnish/Parallel2@12702 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ Macaulay2/d/scclib.c
@@ -724,7 +724,7 @@ char *name;
 int system_acceptBlocking(int so) {
 #if HAVE_SOCKET
   struct sockaddr_in addr;
-  unsigned int addrlen = sizeof addr;
+  size_t addrlen = sizeof addr;
   fcntl(so,F_SETFL,0);
   return accept(so,(struct sockaddr*)&addr,&addrlen);
 #else
